+++ resources/js/useInstrumentationAdvisor.svelte.js
@@ -18,7 +18,7 @@
  *   console.log($advisor.suggestions);   // actionable items
  *
  * @package ZeroBoiler Analytics
- * @version 259.0.0
+ * @version 260.0.0
  */
 
 import { writable, derived } from 'svelte/store';